refactor: 기물의 타입 확인 메서드 추가
